% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Call Leightweight Annotation Widget.}
\usage{
annotate(
  x,
  annotations = annotationstable(),
  width = NULL,
  height = NULL,
  buttons = list(keep = "yellow", drop = "orange"),
  file = NULL,
  ...
)
}
\arguments{
\item{x}{Either a \code{fulltexttable} object or an object that can be
transformed to a \code{fulltexttable} using the \code{fulltexttable()} method.}

\item{annotations}{An optional \code{annotationstable} object. If supplied,
existing annotations will be highlighted in the fulltext displayed.
Defaults to an empty \code{annotationstable} generated by calling
\code{annotationstable()}.}

\item{width}{The width of the annolite htmlwidget.}

\item{height}{The height of the annolite htmlwidget.}

\item{buttons}{To define button/color combinations for annotation mode,
supply a named \code{list} of length-one \code{character} vectors to specify codes
and corresponding colors for highlighting (names are codes, values colors).
Colors that are assigned need to be either valid hex colors or among the
color names R knows about (see \code{grDevices::colors()}). For display mode,
set argument as \code{FALSE}, and there will \emph{not}  be a pop-up menu to create
an annotation if text is selected.}

\item{file}{If a \code{character} vector, a filename to save table with
annotations to disk whenever a new annotation is added. If the filename
ends with ".rds", a RDS file is saved. In all other cases, a csv file
is generated.}

\item{...}{Further arguments passed into call of \code{fulltexttable}.}
}
\value{
A \code{data.frame} with annotations is returned invisibly.
}
\description{
Call Leightweight Annotation Widget.
}
\examples{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
if (interactive()) Y <- annotate(P)
if (interactive()) Y <- annotate(D, annotations = sample_annotation)
}
