% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationstable.R
\name{annotationstable}
\alias{annotationstable}
\alias{is.annotationstable}
\alias{as.annotationstable}
\title{Create annotationstable}
\usage{
annotationstable(
  text = character(),
  code = character(),
  color = character(),
  annotation = character(),
  start = integer(),
  end = integer()
)

is.annotationstable(x)

as.annotationstable(x)
}
\arguments{
\item{text}{A \code{character} vector.}

\item{code}{A \code{character} vector.}

\item{color}{A \code{character} vector.}

\item{annotation}{A \code{character} vector.}

\item{start}{A \code{integer} vector}

\item{end}{integer A \code{integer} vector.}

\item{x}{An object to check whether it is a valid \code{annotationstable} object.}
}
\description{
The annolite package uses an \code{annotationstable} to store and manage the
annotations of an document. An \code{annotationstable} object is a S3
superclass of a \code{data.frame}. An \code{annotationstable} is a
\code{data.frame} with the following columns:
\itemize{
\item \emph{text}: The original text that has been selected to serve as a basis
for an annotation. Note that text selection is not required to cover whole
words. Parts of words may have been selected.
\item \emph{code}: A code that has been assigned to an annotated passage of text.
Available codes need to be predefined before an annotation exercise.
\item \emph{color}: The color used to highlight annotated token sequences. Each
color needs to reflect one code.
\item \emph{annotation}: An explanatory note on the code that has been assigned.
\item \emph{start}: ID of the first token that has been annotated. If the
annotated document is a subcorpus of a CWB corpus, the token ID is the
\code{integer} corpus position.
\item \emph{end}: ID of the last token that has been annotated. If the annotated
document is a subcorpus of a CWB corpus, the token ID is the \code{integer}
corpus position.
}
}
\details{
The function \code{annotationstable()} serves as a constructor for
an \code{annotationstable}. If called without arguments, an empty
\code{annotationstable} is returned. If called with arguments, all input
vectors need to have the same length.

The auxiliary function \code{is.annotationstable()} checks whether
the input objext \code{x} is a \code{annotationstable} object and whether it is
valid. The return value is \code{TRUE} if \code{x} is a valid \code{annotationsstable}
object and \code{FALSE} if not.

The auxiliary function \code{as.annotationstable()} will only accept
objects of class \code{data.frame} and \code{annotationstable} as input \code{x} at this
stage. If \code{x} is a \code{data.frame}, the object will be assigned the class
\code{annotationstable}. If \code{x} is an \code{annotationstable}, it will remain
unchanged. To ensure the validity of the object that is returned, the
object is checked using \code{is.annotationstable()}. The return value is a
valid \code{annotationstable} object.
}
\examples{
empty_annotationstable <- annotationstable()
}
