% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotator.R
\docType{data}
\name{Annotator}
\alias{Annotator}
\title{Annotator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Annotator
}
\description{
object for annotation
}
\section{Fields}{

\describe{
\item{\code{partition}}{a partition object}

\item{\code{type}}{the partition type (e.g. "plpr")}

\item{\code{annotations}}{a JSON string}

\item{\code{filename}}{XXX}

\item{\code{cpos}}{a data.table}
}}

\section{Methods}{

\describe{
  \item{\code{initialize(partition = NULL, filename = NULL)}}{
    if partition is NULL, and filename is provided, restore Annotator-object
    }
  \item{\code{addAnnotation(id, annotation)}}{add an annotation}
  \item{\code{updateAnnotation(id, annotation)}}{update an annotation}
  \item{\code{deleteAnnotation(id)}}{delete an annotation}
}
}

\examples{
\dontrun{
library(polmineR)
library(shiny)
library(shinyjs)
library(R6)
options("polmineR.annoDir" = "/Users/blaette/Lab/tmp/anno")
options("shiny.launch.browser" = browseURL)
options("browser" = "/Applications/Google\\\\ Chrome.app/Contents/MacOS/Google\\\\ Chrome")

filenames <- list.files(getOption("polmineR.annoDir"), full.names = T, patter = "^.*\\\\.RData$")
foo1 <- Annotator$new(filename = "gabriel.json", type = "plpr")
foo1 <- Annotator$new(partition = readRDS(filenames[1]), filename = "gabriel.json")
foo2 <- Annotator$new(partition = readRDS(filenames[2]), filename = "kauder.json")
foo3 <- Annotator$new(partition = readRDS(filenames[3]), filename = "merkel.json")
runApp("/Users/blaette/Lab/gitlab/polmineR.anno/inst/shiny")
}
}
\keyword{datasets}
