% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.annotatordata.R
\name{as.annotatordata}
\alias{as.annotatordata}
\alias{as.annotatordata.plpr_partition}
\alias{as.annotatordata.subcorpus}
\title{Convert object to input for annotator htmlwidget.}
\usage{
as.annotatordata(x, headline)

\method{as.annotatordata}{plpr_partition}(x, headline)

\method{as.annotatordata}{subcorpus}(x, headline)
}
\arguments{
\item{x}{The object to be converted.}

\item{headline}{A headline to append.}
}
\description{
Convert object to input for annotator htmlwidget.
}
\examples{
\dontrun{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.annotatordata(P, headline = "Volker Kauder (CDU)")
annotate(D)
}
}
