% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltextdata.R
\name{as.fulltextdata}
\alias{as.fulltextdata}
\title{Convert object to input for fulltext.}
\usage{
as.fulltextdata(x, headline)
}
\arguments{
\item{x}{The object to be converted.}

\item{headline}{A headline to append.}
}
\description{
Convert object to input for fulltext.
}
\examples{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltextdata(P, headline = "Volker Kauder (CDU)")
fulltext(D)
}
