% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltextlist.R
\name{as.fulltextlist}
\alias{as.fulltextlist}
\alias{as.fulltextlist,subcorpus-method}
\alias{as.fulltextlist,partition-method}
\alias{as.fulltextlist,plpr_subcorpus-method}
\alias{as.fulltextlist,character-method}
\alias{as.fulltextlist,fulltextlist-method}
\alias{as.fulltextlist,FulltextData-method}
\title{Convert object to input for fulltext (table format).}
\usage{
as.fulltextlist(x, ...)

\S4method{as.fulltextlist}{subcorpus}(x, display = "block", name = "")

\S4method{as.fulltextlist}{partition}(x, display = "block", name = "")

\S4method{as.fulltextlist}{plpr_subcorpus}(x, display = "block", name = "")

\S4method{as.fulltextlist}{character}(x, display = "block", tag = "para", name = "", beautify = TRUE)

\S4method{as.fulltextlist}{fulltextlist}(x, ...)

\S4method{as.fulltextlist}{FulltextData}(x, ...)
}
\arguments{
\item{x}{An input object to be converted.}

\item{...}{Further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block"}

\item{name}{An id inserted into tags.}

\item{tag}{A tag.}

\item{beautify}{Remove whitespace before interpunctation.}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
\dontrun{
library(polmineR)
use("polmineR")
x <- corpus("GERMAPARLMINI")
sp <- subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- as.fulltextlist(sp)
}
\dontrun{
x <- corpus("GERMAPARLMINI")
sp <-  subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- as.fulltextlist(sp)
headline <- as.fulltextlist(x = c("Volker", "Kauder", "(CDU)"), tag = "h2")
y2 <- c(headline, y)
}
}
