% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{dialog_radio_buttons}
\alias{dialog_radio_buttons}
\title{Customize bootbox prompt dialog.}
\usage{
dialog_radio_buttons(...)
}
\arguments{
\item{...}{Definition of options in radio button group: Arguments are options, values are colors.}
}
\description{
Customize bootbox prompt dialog.
}
