% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltextlist.R
\name{display}
\alias{display}
\alias{display,fulltextlist-method}
\alias{display<-}
\alias{display<-,fulltextlist-method}
\title{Change display mode of fulltextlist}
\usage{
display(x)

\S4method{display}{fulltextlist}(x)

display(x) <- value

\S4method{display}{fulltextlist}(x) <- value
}
\arguments{
\item{x}{A \code{fulltextlist} object.}

\item{value}{Either "block" or "none".}
}
\description{
Change display mode of fulltextlist
}
\examples{
\dontrun{
li <- corpus("REUTERS") \%>\%  subset(id == "127") \%>\% fulltextlist()
display(li)
display(li) <- "none"
display(li)
}
}
