% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltextlist.R
\name{element}
\alias{element}
\alias{element,fulltextlist-method}
\alias{element<-}
\alias{element<-,fulltextlist-method}
\title{Get and change element tag}
\usage{
element(x)

\S4method{element}{fulltextlist}(x)

element(x) <- value

\S4method{element}{fulltextlist}(x) <- value
}
\arguments{
\item{x}{A \code{fulltextlist} object}

\item{value}{Value to assign to element.}
}
\description{
Get and change element tag
}
\examples{
\dontrun{
li <- corpus("REUTERS") \%>\%  subset(id == "127") \%>\% fulltextlist()
element(li)
element(li) <- "h2"
name(li)
}
}
