% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{emma_chapters_tokenized}
\alias{emma_chapters_tokenized}
\title{Jane Austen's novel Emma (tokenized)}
\format{
A \code{list}
}
\usage{
emma_chapters_tokenized
}
\description{
The novels written by Jane Austen are popular data to illustrate text mining
applications readily available via the R package
\href{https://CRAN.R-project.org/package=janeaustenr}{janeaustenr}. To convey how
\code{annolite} may be used in the context of the tidytext approach (see \href{https://www.tidytextmining.com/}{the book by Julia Silge and David Robinson}), the
package includes a pre-processed object with the chapters of the novel Emma.
}
\seealso{
Other datasets included in the annolite package: 
\code{\link{secretary_general_2000}}
}
\concept{datasets included in the annolite package}
\keyword{datasets}
