% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltext.R
\name{fulltext}
\alias{fulltext}
\alias{fulltext,fulltextlist-method}
\alias{fulltext,SharedData-method}
\alias{fulltext,list-method}
\title{Fulltext output htmlwidget.}
\usage{
fulltext(x, ...)

\S4method{fulltext}{fulltextlist}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = NULL,
  crosstalk = FALSE
)

\S4method{fulltext}{SharedData}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = x$groupName(),
  crosstalk = TRUE
)

\S4method{fulltext}{list}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = "fulltext"
)
}
\arguments{
\item{x}{The data.}

\item{...}{Further arguments to be defined by individual methods.}

\item{annotations}{A table with annotations.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{box}{Logical, whether to put text into a box.}

\item{group}{Name of the crosstalk group, see ....}

\item{crosstalk}{A \code{logial} value, whether to enable crosstalk in
htmlwidget.}
}
\description{
Fulltext output htmlwidget.
}
