% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulltextData.R
\name{FulltextData}
\alias{FulltextData}
\title{Create and Manage Data for fulltext annotation und display widget.}
\description{
Create and Manage Data for fulltext annotation und display widget.

Create and Manage Data for fulltext annotation und display widget.
}
\examples{
library(polmineR)
k <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker == "Volker Kauder") \%>\%
  subset(date == "2009-11-10")
  
F <- FulltextData$new(k, display = "block")
m <- cpos(k, query = "Opposition")
anno <- data.frame(color = "yellow", start = m[,1], end = m[,2])
# F$tooltips(tooltips = list(Opposition = "A", Regierung = "B"))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A fulltexttable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FulltextData$new()}}
\item \href{#method-rename}{\code{FulltextData$rename()}}
\item \href{#method-get_name}{\code{FulltextData$get_name()}}
\item \href{#method-retag}{\code{FulltextData$retag()}}
\item \href{#method-split}{\code{FulltextData$split()}}
\item \href{#method-set_display}{\code{FulltextData$set_display()}}
\item \href{#method-get_display}{\code{FulltextData$get_display()}}
\item \href{#method-clone}{\code{FulltextData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of class \code{CorpusData}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$new(input, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{An object that can be turned into a fulltexttable using \code{as.fulltexttable}.}

\item{\code{...}}{Further arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A class \code{CorpusData} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename"></a>}}
\if{latex}{\out{\hypertarget{method-rename}{}}}
\subsection{Method \code{rename()}}{
Assign new name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$rename(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{New name to assign.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get name(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retag"></a>}}
\if{latex}{\out{\hypertarget{method-retag}{}}}
\subsection{Method \code{retag()}}{
Assign new name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$retag(regex, old, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{regex}}{A regular expression to detect whether a tag should be changed.}

\item{\code{old}}{The old tag.}

\item{\code{new}}{The new tag.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split"></a>}}
\if{latex}{\out{\hypertarget{method-split}{}}}
\subsection{Method \code{split()}}{
Split up table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$split(column = "token", regex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column}}{The column the examine.}

\item{\code{regex}}{A regular expression to detect whether a tag should be changed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_display"></a>}}
\if{latex}{\out{\hypertarget{method-set_display}{}}}
\subsection{Method \code{set_display()}}{
Set visibility status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$set_display(value = c("block", "none"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Visibility status to assign.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_display"></a>}}
\if{latex}{\out{\hypertarget{method-get_display}{}}}
\subsection{Method \code{get_display()}}{
Get visibility status.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$get_display()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FulltextData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
