% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltextlist.R
\name{fulltextlist}
\alias{fulltextlist}
\alias{fulltextlist,subcorpus-method}
\alias{fulltextlist,partition-method}
\alias{fulltextlist,plpr_subcorpus-method}
\alias{fulltextlist,character-method}
\alias{fulltextlist,fulltextlist-method}
\alias{fulltextlist,list-method}
\alias{fulltextlist,data.frame-method}
\alias{c.fulltextlist}
\title{Convert object to input for fulltext (table format).}
\usage{
fulltextlist(x, ...)

\S4method{fulltextlist}{subcorpus}(x, display = "block", name = "")

\S4method{fulltextlist}{partition}(x, display = "block", name = "")

\S4method{fulltextlist}{plpr_subcorpus}(x, display = "block", name = "")

\S4method{fulltextlist}{character}(
  x,
  display = "block",
  element = "para",
  name = "",
  beautify = TRUE
)

\S4method{fulltextlist}{fulltextlist}(x, ...)

\S4method{fulltextlist}{list}(x)

\S4method{fulltextlist}{data.frame}(x, element = "para", display = "block", name = "")

\method{c}{fulltextlist}(...)
}
\arguments{
\item{x}{An input object to be converted.}

\item{...}{Further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block"}

\item{name}{An id inserted into tags.}

\item{element}{An element to assign}

\item{beautify}{Remove whitespace before interpunctation.}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
\dontrun{
x <- corpus("REUTERS") \%>\% subset(id == "127")
fli <- fulltextlist(x)
}
\dontrun{
p <- partition("REUTERS", id = "127")
fli <- fulltextlist(p)
}
\dontrun{
library(polmineR)
use("polmineR")
x <- corpus("GERMAPARLMINI")
sp <- subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltextlist(sp)
}
\dontrun{
x <- corpus("GERMAPARLMINI")
sp <-  subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltextlist(sp)
headline <- fulltextlist(x = c("Volker", "Kauder", "(CDU)"), tag = "h2")
y2 <- c(headline, y)
}
}
