% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltexttable.R
\name{fulltexttable}
\alias{fulltexttable}
\alias{fulltexttable,subcorpus-method}
\alias{fulltexttable,partition-method}
\alias{fulltexttable,plpr_subcorpus-method}
\alias{fulltexttable,character-method}
\alias{fulltexttable,fulltexttable-method}
\alias{fulltexttable,list-method}
\title{Generate fulltexttable}
\usage{
fulltexttable(x, ...)

\S4method{fulltexttable}{subcorpus}(x, display = "block", name = "")

\S4method{fulltexttable}{partition}(x, display = "block", name = "")

\S4method{fulltexttable}{plpr_subcorpus}(x, display = "block", name = "")

\S4method{fulltexttable}{character}(x, display = "block", element = "para", name = "")

\S4method{fulltexttable}{fulltexttable}(x, display = "block", element = "para", name = "")

\S4method{fulltexttable}{list}(x, display = "block", element = "para", name = "")
}
\arguments{
\item{x}{Object to be converted to \code{fulltexttable}. Can be a \code{character}
vector with tokens of a chunk of text, or a \code{list} of \code{character} vectors
representing chunks. Alternatively, objects used in the polmineR package
can be processed (\code{subcorpus}, \code{partition}, \code{plpr_subcorpus}).}

\item{...}{Further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block" by default so that text is displayed.}

\item{name}{A lenght-one \code{character} vector assigned to column "name" of
\code{fulltexttable}. Used as name attribute of elements wrapping chunks of text
to make selections between different documents. Defaults to "" as it will
not be used unless crosstalk is used to skip through documents.}

\item{element}{An element to assign}
}
\description{
The \code{fulltexttable} class is a superclass of the traditional \code{data.frame} to
represent the information to show the fulltext of a document (or documents)
as layouted HTML. Rows of a \code{fulltexttable} represent chunks of text
(paragraphs, headlines, blockquotes) that are formatted according to
information defined by the columns of the table. A \code{fulltexttable} has the
following columns:
\itemize{
\item \emph{name}: A \code{character} vector, serves as an identifier to distinguish the
content of different documents. Defaults to "", but required to be defined
when different documents are to be combined in one \code{fulltexttable} and
crosstalk is used to filter, or to select documents for display.
\item \emph{element}: The HTML element that wraps the tokens of a chunk of text that
are formatted according CSS instructions. For ordinary paragraphs, the
element 'para' should be used rathen than 'p' to avoid conclicts between
(potentially) CSSs.
\item \emph{style}: Content of the style attribute of the element. Used at this stage
to set the display mode. Defaults to "block" (chunk is displayed), but needs
to be "display:none" when crossstalk is used to filter documents, or to make
selections.
\item \emph{tokenstream}: A list of three-column \code{data.frame} objects that are nested
into the cells of the \code{fulltexttable}. The HTML for fulltext display will be
generated from this data upon calling the \emph{annolite} HTML widget. Keeping the
fulltext data in the nested \code{data.frame} ensures that data can be changed at
all times.
The \code{fulltexttable()} method is used to construct a \code{fulltexttable} from
different input objects.
}
}
\examples{
library(polmineR)
use("polmineR")

x <- corpus("REUTERS") \%>\% subset(id == "127")
tbl <- fulltexttable(x)
p <- partition("REUTERS", id = "127")
fli <- fulltexttable(p)
library(polmineR)
use("polmineR")

x <- corpus("GERMAPARLMINI")
sp <- subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltexttable(sp)
x <- corpus("GERMAPARLMINI")
sp <-  subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltexttable(sp)
headline <- fulltexttable(x = c("Volker", "Kauder", "(CDU)"), element = "h2")
y2 <- c(headline, y)
}
