% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{sample_annotation}
\alias{sample_annotation}
\title{Sample annotation of a speech given by Volker Kauder}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.
}
\usage{
sample_annotation
}
\description{
Sample annotation of a speech given by Volker Kauder
}
\keyword{datasets}
