% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltextlist.R
\name{name<-,fulltextlist-method}
\alias{name<-,fulltextlist-method}
\alias{name,fulltextlist-method}
\title{Manage names of fulltextlist}
\usage{
\S4method{name}{fulltextlist}(x) <- value

\S4method{name}{fulltextlist}(x)
}
\arguments{
\item{x}{A \code{fulltextlist} object.}

\item{value}{The value to assign.}
}
\description{
Manage names of fulltextlist
}
\examples{
\dontrun{
li <- corpus("REUTERS") \%>\%  subset(id == "127") \%>\% fulltextlist()
name(li)
name(li) <- "reuters127"
name(li)
}
}
