% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{secretary_general_2000_speech}
\alias{secretary_general_2000_speech}
\alias{secretary_general_2000_annotations}
\title{Annotated speech of Kofi Annan}
\format{
The object \code{secretary_general_2000_speech} is a \code{fulltexttable}, see
documentation object \code{\link{fulltexttable}} for an explanation of the
data structure. To inspect the code that has been used to prepare the
\code{fulltexttable}, see the file 'annan2000.R' in the data-raw subdirectory of
the package.

The object \code{secretary_general_2000_annotations} is a
\code{annotationstable}, see documentation obje \code{\link{annotationstable}}
for an explanation of the data structure. To inspect the code that has been
used to prepare the \code{annotationstable}, see the file 'annan2000.R' in the
data-raw subdirectory of the package.
}
\usage{
secretary_general_2000_speech

secretary_general_2000_annotations
}
\description{
The package includes the fulltext of Kofi Annan's speech opening the 2000
Millennium Summit (object \code{secretary_general_2000_speech}) and a sample
annotation of the speech (object \code{secretary_general_2000_annotations}).
}
\keyword{datasets}
