% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\name{annoliteOutput}
\alias{annoliteOutput}
\alias{renderAnnolite}
\title{Render annolite htmlwidget in shiny apps.}
\usage{
annoliteOutput(outputId, width = "100\%", height = "100\%")

renderAnnolite(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read the value from.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{expr}{An expression (...).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if
you want to save an expression in a variable.}
}
\description{
Render annolite htmlwidget in shiny apps.
}
