% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltextlist.R
\name{split,fulltextlist-method}
\alias{split,fulltextlist-method}
\title{Split up fulltextlist}
\usage{
\S4method{split}{fulltextlist}(x, regex)
}
\arguments{
\item{x}{A \code{fulltextlist} object}

\item{regex}{A regular expression.}
}
\description{
Split up fulltextlist
}
