% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{secretary_general_2000}
\alias{secretary_general_2000}
\alias{unga_migrationspeeches_fulltext}
\alias{unga_migrationspeeches_anntationstable}
\title{Sample documents and annotations from the UN General Assembly}
\format{
The object \code{secretary_general_2000} is a \code{fulltexttable}.

The object \code{unga_migrationspeechess_fulltext} is a \code{fulltexttable}

The object \code{unga_migrationspeeches_annotationstable} is an
\code{annotationstable}.
}
\usage{
secretary_general_2000

unga_migrationspeeches_fulltext

unga_migrationspeeches_anntationstable
}
\description{
The package includes three pieces of data derived from the corpus of speeches
given in the UN General Assembly (UNGA corpus, \href{https://zenodo.org/record/3831472#.X5nMp1kxlZ0}{available at Zenodo}).
}
\details{
The file
\href{https://github.com/PolMine/annolite/blob/master/data-raw/unga_topicmodelling.R}{unga_topicmodelling.R}
in the data-raw folder of the package has been used to create the data
objects.
}
\seealso{
Other datasets included in the annolite package: 
\code{\link{emma_chapters_tokenized}}
}
\concept{datasets included in the annolite package}
\keyword{datasets}
