% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{unga_migrationspeeches_fulltext}
\alias{unga_migrationspeeches_fulltext}
\alias{unga_migrationspeeches_anntationstable}
\title{Speeches on migration in the UN General Assembly}
\format{
The object \code{unga_migrationspeeches_fulltext} is a \code{fulltexttable},
see the documentation of the function \code{fulltexttable} for an explanation
of the data structure of the class \code{fulltexttable}.

The object \code{unga_migrationspeeches_anntationstable} is a
\code{anntationstable}, see the documentation of the function \code{anntationstable}
for an explanation of the data structure of the class \code{anntationstable}.
}
\usage{
unga_migrationspeeches_fulltext

unga_migrationspeeches_anntationstable
}
\description{
To illustrate how the \emph{annolite} HTML widget can be used for evaluating a
topic model, i.e. for inspecting documents where a topic is present among the
m first topics, the package includes a set of UN General Assembly speeches
(presumably) addressing migration affairs (object
\code{unga_migrationspeeches_fulltext}) and an \code{annotationstable} (object
\code{unga_migrationspeeches_anntationstable}) to highlight tokens indicative of
the migration topic.
}
\details{
To inspect the code that has been used to prepare the object, see the file
'unga_topicmodelling.R' in the data-raw subdirectory of the package.
}
\keyword{datasets}
