% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{package}
\name{annolite}
\alias{annolite}
\alias{annolite-package}
\alias{sample_annotation}
\alias{emma_chapters_tokenized}
\alias{secretary_general_2000}
\title{Annolite htmlwidget.}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.

An object of class \code{list} of length 55.

An object of class \code{fulltextlist} (inherits from \code{list}, \code{vector}) of length 21.
}
\usage{
annolite(data, width = NULL, height = NULL, dialog = NULL, box = TRUE)

sample_annotation

emma_chapters_tokenized

secretary_general_2000
}
\arguments{
\item{data}{The data that is passed to the JavaScript that generates the output. Expected to be
a list of lists that provide data on sections of text. Each of the sub-lists is to be a named
list of a character vector with the HTML element the section will be wrapped into, and
a \code{data.frame} (or a list) with a column "token", and a column "id".}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{dialog}{The dialog.}

\item{box}{Logical, whether to put text into a box.}
}
\description{
Annolite htmlwidget.
}
\examples{
library(polmineR)
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
y <- annolite(
  data = list(paragraphs = fulltextlist(P), annotations = annotationstable()),
  width = "100\%",
  dialog = list(choices = dialog_radio_buttons(keep = "yellow", drop = "orange"))
)
}
\author{
Andreas Blaette
}
\keyword{datasets}
