% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{package}
\name{annolite}
\alias{annolite}
\alias{annolite-package}
\alias{sample_annotation}
\title{R-Package 'annolite'.}
\format{
An object of class \code{data.frame} with 4 rows and 5 columns.
}
\usage{
sample_annotation
}
\description{
Leightweight fulltext display and annotation HTML widget.
}
\author{
Andreas Blaette
}
\keyword{datasets}
