% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{package}
\name{annolite}
\alias{annolite}
\alias{annolite-package}
\alias{annolite,fulltexttable-method}
\alias{annolite,SharedData-method}
\alias{sample_annotation}
\alias{emma_chapters_tokenized}
\alias{secretary_general_2000}
\alias{unga_migrationspeeches_fulltext}
\alias{unga_migrationspeeches_anntationstable}
\title{Instantiate annolite htmlwidget.}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.

An object of class \code{list} of length 55.

An object of class \code{fulltexttable} (inherits from \code{data.frame}, \code{list}, \code{oldClass}, \code{vector}) with 21 rows and 4 columns.

An object of class \code{fulltexttable} (inherits from \code{data.frame}, \code{list}, \code{oldClass}, \code{vector}) with 178 rows and 4 columns.

An object of class \code{annotationstable} (inherits from \code{data.frame}, \code{list}, \code{oldClass}, \code{vector}) with 2041 rows and 6 columns.
}
\usage{
annolite(x, ...)

\S4method{annolite}{fulltexttable}(
  x,
  annotations = annotationstable(),
  buttons = list(keep = "yellow", drop = "lightgreen"),
  width = "100\%",
  height = NULL,
  box = TRUE,
  crosstalk = FALSE,
  layout = "filter",
  group = "fulltext"
)

\S4method{annolite}{SharedData}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = x$groupName(),
  crosstalk = TRUE
)

sample_annotation

emma_chapters_tokenized

secretary_general_2000

unga_migrationspeeches_fulltext

unga_migrationspeeches_anntationstable
}
\arguments{
\item{x}{A \code{fulltexttable}.}

\item{...}{Further arguments}

\item{annotations}{A \code{annotationstable}.}

\item{buttons}{Definition of buttons (named list).}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{box}{Logical, whether to put text into a box.}

\item{crosstalk}{\code{logical}, whether to use crosstalk}

\item{layout}{Relevant for crosstalk mode}

\item{group}{crosstalk group}
}
\description{
Instantiate annolite htmlwidget.
}
\examples{
library(polmineR)
sc <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker == "Volker Kauder" & date == "2009-11-10")
tab <- fulltexttable(sc)
y <- annolite(
  x = tab, annotations = annotationstable(),
  width = "100\%",
  buttons = list(keep = "yellow", drop = "orange")
)
}
\author{
Andreas Blaette
}
\keyword{datasets}
