% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{package}
\name{annolite}
\alias{annolite}
\alias{annolite-package}
\alias{annolite,fulltexttable-method}
\alias{annolite,SharedData-method}
\title{Instantiate annolite htmlwidget}
\usage{
annolite(x, ...)

\S4method{annolite}{fulltexttable}(
  x,
  annotations = annotationstable(),
  buttons = list(keep = "yellow", drop = "lightgreen"),
  width = "100\%",
  height = NULL,
  box = TRUE,
  crosstalk = FALSE,
  layout = "filter",
  group = "fulltext"
)

\S4method{annolite}{SharedData}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = x$groupName(),
  crosstalk = TRUE
)
}
\arguments{
\item{x}{Either a \code{fulltexttable} object or an object that can be
transformed to a \code{fulltexttable} using the \code{fulltexttable()} method.}

\item{...}{Further arguments}

\item{annotations}{An optional \code{annotationstable} object. If supplied,
existing annotations will be highlighted in the fulltext displayed.
Defaults to an empty \code{annotationstable} generated by calling
\code{annotationstable()}.}

\item{buttons}{To define button/color combinations for annotation mode,
supply a named \code{list} of length-one \code{character} vectors to specify codes
and corresponding colors for highlighting (names are codes, values colors).
Colors that are assigned need to be either valid hex colors or among the
color names R knows about (see \code{grDevices::colors()}). For display mode,
set argument as \code{FALSE}, and there will \emph{not}  be a pop-up menu to create
an annotation if text is selected.}

\item{width}{The width of the annolite htmlwidget.}

\item{height}{The height of the annolite htmlwidget.}

\item{box}{Length-one \code{logical} value, whether draw box around HTML widget
with fulltext display.}

\item{crosstalk}{\code{logical}, whether to use crosstalk}

\item{layout}{Relevant for crosstalk mode}

\item{group}{An identifier for a Crosstalk group. HTML widgets within one
Crosstalk group can communicate with each other, see documentation of the
argument \code{group} of \code{crosstalk::SharedData}. Defaults to "fulltext".}
}
\description{
The HTML widget \emph{annolite} is the core of the package and the \code{annolite()}
method is the constructor to generate it. The widget is designed for two
basic uses: (a) The widget's annotation mode offers a pure R workflow for
basic text annotation tasks. In this scenario, it is used within a Shiny
Gadget called via \code{annotate()}. (b) The widget's display mode is designed to
inspect the fulltext of a document with highlighted annotations either in an
interactive R session or embedded in Rmarkdown documents. The widget supports
crosstalk, enabling the interactive selection among a set of documents within
an HTML document generated from Rmarkdown.
}
\examples{
library(polmineR)
sc <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker == "Volker Kauder" & date == "2009-11-10")
tab <- fulltexttable(sc)
y <- annolite(
  x = tab, annotations = annotationstable(),
  width = "100\%",
  buttons = list(keep = "yellow", drop = "orange")
)
}
\author{
Andreas Blaette
}
