% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Leightweight Annotation Widget.}
\usage{
annotate(input, width = NULL, height = NULL, codes = c(keep = "green",
  drop = "orange", reconsider = "grey"))
}
\arguments{
\item{input}{A list of data.frames according to the requirements of fulltext
input.}
}
\description{
Leightweight Annotation Widget.
}
\examples{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltextdata(P, headline = "Volker Kauder (CDU)")
if (interactive()) Y <- annotate(D)
}
