% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Shiny Gadget for Text Annotation.}
\usage{
annotate(
  x,
  annotations = annotationstable(),
  width = NULL,
  height = NULL,
  buttons = list(keep = "yellow", drop = "orange"),
  file = NULL,
  ...
)
}
\arguments{
\item{x}{Either a \code{fulltexttable} object or an object that can be
transformed to a \code{fulltexttable} using the \code{fulltexttable()} method.}

\item{annotations}{An optional \code{annotationstable} object. If supplied,
existing annotations will be highlighted in the fulltext displayed.
Defaults to an empty \code{annotationstable} generated by calling
\code{annotationstable()}. If \code{annotations} is a \code{data.frame}, it will be turned
into an \code{annotationstable} using \code{as.annotationstable()}; the validity of
the object will be checked using \code{is.annotationstable()}.}

\item{width}{The width of the annolite htmlwidget.}

\item{height}{The height of the annolite htmlwidget.}

\item{buttons}{To define button/color combinations for annotation mode,
supply a named \code{list} of length-one \code{character} vectors to specify codes
and corresponding colors for highlighting (names are codes, values colors).
Colors that are assigned need to be either valid hex colors or among the
color names R knows about (see \code{grDevices::colors()}). For display mode,
set argument as \code{FALSE}, and there will \emph{not}  be a pop-up menu to create
an annotation if text is selected.}

\item{file}{If a \code{character} vector, a filename that is used to save table
with annotations to disk whenever a new annotation is added. If the
filename ends with ".rds", a RDS file is saved. In all other cases, a csv
file is generated.  If the parent directory of the file does not exists,
\code{annotate()} will abort issuing an error message. Note that existing files
will be overwritten. If argument is \code{NULL} (default), no file with
annotations will be generated an upated.}

\item{...}{Further arguments passed into call of \code{fulltexttable()}.}
}
\value{
A \code{data.frame} with annotations (class \code{\link{annotationstable}}) is
returned invisibly.
}
\description{
Calling the method \code{annotate()} on an object that is either a \code{fulltexttable}
or that can be transformed to a \code{fulltexttable} (using the method
\code{fulltexttable()}) will launch a Shiny Widget as a leightweight text
annotation tool. The HTML widget \code{annolite} is the core of the gadget. It
provides the essential functionality for highlighter-and-pencil-style
annotation. Wrapping the \emph{annolite}  HTML wideget in a Shiny Gadget
facilitates the implementation of a pure R workflow for generating and
procesing text annotations.
}
\details{
The HTML widget will permit the following actions:
\itemize{
\item When selecting a range of text, a modal dialogue will open to create a new
annotation; using the buttons that have been defined (argument \code{buttons}),
you assign an annotation. Optionally, you can write an explanatory note on
an annotation.
\item When you click on an existing annotation, you can delete the annotation.
Modifying an existing annotation is not possible yet.
\item When you select a range of text including more than one word and that includes
an existing annotation, a modal dialogue will inform you that deleting an
existing annotation will require selecting only one word.
}
}
\examples{
if (interactive()){
y <- annotate(
  x = secretary_general_2000_speech,
  annotations = annolite::secretary_general_2000_annotations
)
}
}
