% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator_widget.R
\name{annotate}
\alias{annotate}
\title{Leightweight Annotation Widget.}
\usage{
annotate(input, freetext = FALSE, codes = c(great = "lightgreen", positive =
  "green", neutral = "darkgrey", negative = "red", obnoxious = "darkred"))
}
\arguments{
\item{input}{A list of data.frames according to the requirements of fulltext
input.}

\item{freetext}{Logical, whether to include a freetext field.}

\item{codes}{A named character vector with codes to assign (names of the
vector), and colors for highlighting.}
}
\description{
Leightweight Annotation Widget.
}
\examples{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltextdata(P, headline = "Volker Kauder (CDU)")
if (interactive()) Y <- annotate(D)
}
