% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Leightweight Annotation Widget.}
\usage{
annotate(
  input,
  width = NULL,
  height = NULL,
  dialog = list(choices = dialog_radio_buttons(keep = "yellow", drop = "orange"),
    callback = dialog_default_callback())
)
}
\arguments{
\item{input}{A list of data.frames according to the requirements of fulltext
input.}

\item{width}{The width of the fulltext htmlwidget.}

\item{height}{The height of the fulltext htmlwidget.}

\item{dialog}{Specification of the dialog box.}
}
\description{
Leightweight Annotation Widget.
}
\examples{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltextdata(P, headline = "Volker Kauder (CDU)")
D[["annotations"]] <- sample_annotation
if (interactive()) Y <- annotate(D)
}
