% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationstable.R
\name{annotationstable}
\alias{annotationstable}
\title{Create empty annotationstable}
\usage{
annotationstable(
  text = character(),
  code = character(),
  color = character(),
  annotation = character(),
  start = integer(),
  end = integer()
)
}
\arguments{
\item{text}{character}

\item{code}{character}

\item{color}{character}

\item{annotation}{character}

\item{start}{integer}

\item{end}{integer}
}
\description{
Create empty annotationstable
}
