% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fulltextdata.R
\name{as.fulltextdata}
\alias{as.fulltextdata}
\alias{as.fulltextdata.plpr_partition}
\alias{as.fulltextdata.subcorpus}
\title{Convert object to input for fulltext.}
\usage{
as.fulltextdata(x, headline)

\method{as.fulltextdata}{plpr_partition}(x, headline)

\method{as.fulltextdata}{subcorpus}(x, headline)
}
\arguments{
\item{x}{The object to be converted.}

\item{headline}{A headline to append.}
}
\description{
Convert object to input for fulltext.
}
\examples{
\dontrun{
library(polmineR)
use("polmineR")
P <- partition("GERMAPARLMINI", speaker = "Volker Kauder", date = "2009-11-10")
D <- as.fulltextdata(P, headline = "Volker Kauder (CDU)")
fulltext(D)
}
}
