% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{dialog_default_callback}
\alias{dialog_default_callback}
\alias{dialog_radio_buttons}
\title{Customization for bootbox prompt dialog.}
\usage{
dialog_default_callback()

dialog_radio_buttons(...)
}
\arguments{
\item{...}{Definition of options in radio button group: Arguments are options, values are colors.}
}
\description{
Customization for bootbox prompt dialog.
}
