% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annolite.R
\docType{data}
\name{emma_chapters_tokenized}
\alias{emma_chapters_tokenized}
\title{Jane Austen's Emma}
\format{
A list of lists of character vectors. Each list of character vectors
represents a chapter of Emma.
}
\usage{
emma_chapters_tokenized
}
\description{
For explaining how a \code{fulltexttable} can be generated from any kind
of textual data, the package includes the tokenized text of the first
five chapters of Jane Austen's novel \emph{Emma}.
}
\details{
To inspect the code that has been used to prepare the object, see the file
'emma.R' in the data-raw subdirectory of the package.
}
\keyword{datasets}
