% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltext.R
\name{fulltext}
\alias{fulltext}
\title{Fulltext output.}
\usage{
fulltext(data = list(token = LETTERS[1:10], cpos = 1:10), width = NULL,
  height = NULL, codes = c(keep = "green", drop = "orange", reconsider =
  "grey"))
}
\arguments{
\item{data}{The data that is passed to the JavaScript that generates the output. Expected to be
a list of lists that provide data on sections of text. Each of the sub-lists is to be a named
list of a character vector with the HTML element the section will be wrapped into, and 
a \code{data.frame} (or a list) with a column "token", and a column "cpos".}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}
}
\description{
Fulltext output.
}
