% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltext.R
\name{fulltext}
\alias{fulltext}
\title{Fulltext output.}
\usage{
fulltext(tokendata_file = system.file(package = "fulltext", "extdata", "json",
  "abc.json"), width = NULL, height = NULL)
}
\arguments{
\item{tokendata_file}{xxx}

\item{width}{xxx}

\item{height}{xxx}
}
\description{
Fulltext output.
}
