% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltext.R
\name{fulltext}
\alias{fulltext}
\alias{fulltext.data.frame}
\alias{fulltext.SharedData}
\alias{fulltext.FulltextData}
\alias{fulltext.list}
\title{Fulltext output htmlwidget.}
\usage{
fulltext(x, width = NULL, height = NULL, box = TRUE, group = NULL)

\method{fulltext}{data.frame}(x, width = "100\%", height = NULL, box = TRUE, group = NULL)

\method{fulltext}{SharedData}(x, width = "100\%", height = NULL, box = TRUE, group = x$groupName())

\method{fulltext}{FulltextData}(x, width = "100\%", height = NULL, box = TRUE, group = NULL)

\method{fulltext}{list}(x, width = "100\%", height = NULL, box = TRUE, group = "fulltext")
}
\arguments{
\item{x}{The data.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{box}{Logical, whether to put text into a box.}

\item{group}{Name of the crosstalk group, see ....}
}
\description{
Fulltext output htmlwidget.
}
