% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltext.R
\name{fulltext}
\alias{fulltext}
\alias{fulltext.fulltextlist}
\alias{fulltext.SharedData}
\alias{fulltext.FulltextData}
\alias{fulltext.list}
\title{Fulltext output htmlwidget.}
\usage{
fulltext(x, annotations, width = NULL, height = NULL, box = TRUE, group = NULL)

\method{fulltext}{fulltextlist}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = NULL
)

\method{fulltext}{SharedData}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = x$groupName()
)

\method{fulltext}{FulltextData}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = NULL
)

\method{fulltext}{list}(
  x,
  annotations = NULL,
  width = "100\%",
  height = NULL,
  box = TRUE,
  group = "fulltext"
)
}
\arguments{
\item{x}{The data.}

\item{annotations}{A table with annotations.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{box}{Logical, whether to put text into a box.}

\item{group}{Name of the crosstalk group, see ....}
}
\description{
Fulltext output htmlwidget.
}
