% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulltexttable.R
\name{fulltexttable}
\alias{fulltexttable}
\alias{fulltexttable,subcorpus-method}
\alias{fulltexttable,partition-method}
\alias{fulltexttable,plpr_subcorpus-method}
\alias{fulltexttable,character-method}
\alias{fulltexttable,fulltexttable-method}
\alias{fulltexttable,list-method}
\title{Convert object to input for fulltext (table format).}
\usage{
fulltexttable(x, ...)

\S4method{fulltexttable}{subcorpus}(x, display = "block", name = "")

\S4method{fulltexttable}{partition}(x, display = "block", name = "")

\S4method{fulltexttable}{plpr_subcorpus}(x, display = "block", name = "")

\S4method{fulltexttable}{character}(x, display = "block", element = "para", name = "")

\S4method{fulltexttable}{fulltexttable}(x, display = "block", element = "para", name = "")

\S4method{fulltexttable}{list}(x, display = "block", element = "para", name = "")
}
\arguments{
\item{x}{An input object to be converted.}

\item{...}{Further arguments defined by methods.}

\item{display}{The initial value of the html style argument. Either "block"
or "none". Should usually be "block"}

\item{name}{An id inserted into tags.}

\item{element}{An element to assign}
}
\description{
Convert object to input for fulltext (table format).
}
\examples{
\dontrun{
x <- corpus("REUTERS") \%>\% subset(id == "127")
fli <- fulltexttable(x)
}
\dontrun{
p <- partition("REUTERS", id = "127")
fli <- fulltexttable(p)
}
\dontrun{
library(polmineR)
use("polmineR")
x <- corpus("GERMAPARLMINI")
sp <- subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltexttable(sp)
}
\dontrun{
x <- corpus("GERMAPARLMINI")
sp <-  subset(x, speaker == "Volker Kauder" & date == "2009-11-10")
y <- fulltexttable(sp)
headline <- fulltexttable(x = c("Volker", "Kauder", "(CDU)"), element = "h2")
y2 <- c(headline, y)
}
}
