% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{annotatorOutput}
\alias{annotatorOutput}
\alias{renderAnnotator}
\title{Render annotator htmlwidget in shiny apps.}
\usage{
annotatorOutput(outputId, width = "100\%", height = "100\%")

renderAnnotator(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read the value from.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{expr}{An expression (...).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if
you want to save an expression in a variable.}
}
\description{
Render annotator htmlwidget in shiny apps.
}
